inference_train/requirements.txt
argcomplete==1.10.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
compressed-rtf==1.0.6
cssselect2==0.4.1
cycler==0.11.0
dataclasses==0.8
defusedxml==0.7.1
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
filelock==3.4.1
h5py==3.1.0
html5lib==1.1
huggingface-hub==0.4.0
idna==3.3
IMAPClient==2.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
lxml==4.9.1
matplotlib==3.3.4
mplcursors==0.5.2
numpy==1.19.5
olefile==0.46
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
pdfminer.six==20191110
Pillow==8.4.0
pyclipper==1.3.0.post3
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.7
PyPDF2==2.10.2
pypdfium2==2.11.0
pyphen==0.11.0
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==2.11.1
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
semantic-version==2.9.0
setuptools-rust==1.1.2
Shapely==1.8.5.post1
six==1.12.0
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
textract==1.6.5
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.11.5
torch==1.10.1
torchinfo==1.5.4
torchvision==0.11.2
tqdm==4.62.3
transformers==4.16.2
typing_extensions==4.1.1
tzdata==2022.6
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.12
WeasyPrint==52.5
webencodings==0.5.1
xlrd==1.2.0
XlsxWriter==3.0.3
zipp==3.6.0
python-magic==0.4.25
